+++ inc/test.h
@@ -10,10 +10,10 @@
 struct test_table
 {
 
-	uint32_t start;
-	uint32_t end;
-	uint32_t in;
-	uint32_t out;
+	int start;
+	int end;
+	int in;
+	int out;
 
 };
 
@@ -30,31 +30,31 @@ void Test_init( volatile test_table_t *t );
 
 /******** Test_start *********
 * Records a starting value to a test table.
-*  Inputs: A test_table struct, uint32_t start time
+*  Inputs: A test_table struct, int start time
 * Outputs: none
 */
 void Test_start( volatile test_table_t *t );
 
 /******** Test_end *********
 * Records an ending value to a test table.
-*  Inputs: A test_table struct, uint32_t end time
+*  Inputs: A test_table struct, int end time
 * Outputs: none
 */
 void Test_end( volatile test_table_t *t );
 
 /******** Test_in *********
 * Records an inputted value to a test table.
-*  Inputs: A test_table struct, uint32_t in value
+*  Inputs: A test_table struct, int in value
 * Outputs: none
 */
-void Test_in( volatile test_table_t *t, uint32_t in );
+void Test_in( volatile test_table_t *t, int in );
 
 /******** Test_out *********
 * Records an outputted value to a test table.
-*  Inputs: A test_table struct, uint32_t out value
+*  Inputs: A test_table struct, int out value
 * Outputs: none
 */
-void Test_out( volatile test_table_t *t, uint32_t out );
+void Test_out( volatile test_table_t *t, int out );
 
 /******** Test_calculate *********
 * Calculates the differences between values recorded in a test table,
@@ -68,7 +68,7 @@ void Test_calculate( volatile test_table_t *t );
 // Adds arbitrary number of bytes to the UART transmission buffer.
 //  Inputs: pointer to a contiguous block of data, the number of bytes
 // Outputs: none
-extern void Uart_send( volatile void* data, uint16_t length );
+extern void Uart_send( volatile void* data, int length );
 
 #define TEST_H_ 1
 #endif
